Multi metric formatted example data

A list object containing simulated metric data for package examples. This example is for a multi metric indicator (a vector of random numbers). This has been prepped using the data_prep function, for use with IEAnalyzeR's plotting function.

Usage

multi_data_formatted

Format

A data frame with 132 rows and 5 variables:

date

A vector of months.

temp_a

An vector of random numbers.

salinity_a

A vector of random numbers.

temp_b

An vector of random numbers.

salinity_b

A vector of random numbers.

Source

Simulated data